# Backend/api/user.swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "ClinicCare Pro User API",
    "version": "1.0.0",
    "description": "Swagger-style specification for user CRUD and user management endpoints."
  },
  "servers": [
    {
      "url": "http://localhost:5000"
    }
  ],
  "paths": {
    "/api/users": {
      "get": {
        "summary": "List all users",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "User list loaded"
          }
        }
      },
      "post": {
        "summary": "Create a new user",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully"
          }
        }
      }
    },
    "/api/users/{id}": {
      "get": {
        "summary": "Get a single user by id",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User loaded successfully"
          }
        }
      },
      "put": {
        "summary": "Update a user",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User updated successfully"
          }
        }
      },
      "delete": {
        "summary": "Delete a user",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User deleted successfully"
          }
        }
      }
    },
    "/api/users/{id}/block": {
      "put": {
        "summary": "Block or unblock a user",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "blocked": {
                    "type": "boolean"
                  }
                },
                "required": [
                  "blocked"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User block status updated"
          }
        }
      }
    },
    "/api/users/swagger.json": {
      "get": {
        "summary": "Get Swagger-style user API spec",
        "responses": {
          "200": {
            "description": "Swagger JSON returned"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "CreateUserRequest": {
        "type": "object",
        "required": [
          "name",
          "email",
          "phone",
          "role",
          "password"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "phone": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "example": "doctor"
          },
          "password": {
            "type": "string",
            "format": "password"
          },
          "specialization": {
            "type": "string"
          }
        }
      },
      "UpdateUserRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "phone": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "password": {
            "type": "string",
            "format": "password"
          },
          "specialization": {
            "type": "string"
          },
          "blocked": {
            "type": "boolean"
          }
        }
      }
    }
  }
}